+++ grid.h
@@ -3,15 +3,15 @@
 
 #include <vector>
 
-std::vector <double> calcVolume(std::vector <double> S, std::vector <double> dx);
+std::vector <long double> calcVolume(std::vector <long double> S, std::vector <long double> dx);
 
-std::vector <double> evalX(double a, double b);
+std::vector <long double> evalX(long double a, long double b);
 
-std::vector <double> evalDx(std::vector <double> x);
+std::vector <long double> evalDx(std::vector <long double> x);
 
-std::vector <double> evalS(std::vector <double> geom,
-                           std::vector <double> x,
-                           std::vector <double> dx);
+std::vector <long double> evalS(std::vector <long double> geom,
+                           std::vector <long double> x,
+                           std::vector <long double> dx);
