+++ optimizer.h
@@ -1,7 +1,7 @@
 #ifndef optimizer_h
 #define optimizer_h
 
-void design(std::vector <double> x, std::vector <double> dx,
-            std::vector <double> S, std::vector <double> designVar);
+void design(std::vector <long double> x, std::vector <long double> dx,
+            std::vector <long double> S, std::vector <long double> designVar);
 
 #endif
